Non-exhaustive enums: drop '@nonexhaustive' to just have '@exhaustive'
We may need this temporarily to implement the overlays, but we don't
really want to make the language more complicated in this way. The
only +/- attribute pair we have today is @objc/@nonobjc, and that
one's actually meaningful.

Also, discourage the "member" approach even more. The core team is
also not in favor, so there's not much point in that being discussed
during the formal review period.